requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
python-multipart
fastapi_users
fastapi_users_db_sqlalchemy
python-dotenv
cloudinary
requests
